move direct logging api under apis folder
