+++ apis/logging/v1beta1/register.go
